LayoutTests/imported/w3c/web-platform-tests/css/cssom/CSSKeyframesRule-expected.txt


FAIL name, cssRules, appendRule, findRule, deleteRule assert_equals: CSSKeyframesRule cssText attribute with CSS-wide keyword name expected "@keyframes\"initial\"{}" but got "@keyframesinitial{}"
PASS indexed getter, length
